libafl-fuzz: move each feedback to it's own file
