+++ fuzzers/libafl-fuzz/src/fuzzer.rs
@@ -20,7 +20,7 @@ use libafl::{
         StdPowerMutationalStage,
     },
     state::{HasCorpus, HasCurrentTestcase, HasStartTime, StdState},
-    Error, HasFeedback, HasMetadata,
+    Error, HasMetadata, HasFeedback,
 };
 use libafl_bolts::{
     current_nanos, current_time,
@@ -37,7 +37,7 @@ use serde::{Deserialize, Serialize};
 use crate::{
     afl_stats::AflStatsStage,
     corpus::{set_corpus_filepath, set_solution_filepath},
-    feedback::{CustomFilepathToTestcaseFeedback, SeedFeedback},
+    feedback::{filepath::CustomFilepathToTestcaseFeedback, seed::SeedFeedback},
     run_fuzzer_with_stage, Opt, AFL_DEFAULT_INPUT_LEN_MAX, AFL_DEFAULT_INPUT_LEN_MIN,
     SHMEM_ENV_VAR,
 };
